// src/api-client/models/participant-info.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Spring-Adam-Store
 * Api Service
 *
 * The version of the OpenAPI document: v1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface ParticipantInfo
 */
export interface ParticipantInfo {
    /**
     * 
     * @type {number}
     * @memberof ParticipantInfo
     */
    'userId'?: number;
    /**
     * 
     * @type {string}
     * @memberof ParticipantInfo
     */
    'email'?: string;
    /**
     * 
     * @type {string}
     * @memberof ParticipantInfo
     */
    'name'?: string;
    /**
     * 
     * @type {string}
     * @memberof ParticipantInfo
     */
    'avatarUrl'?: string;
}